move quotient ops to file
