+++ crates/prover/src/core/backend/gpu/accumulation.rs
@@ -1,6 +1,6 @@
-use crate::core::{air::accumulation::AccumulationOps, fields::secure_column::SecureColumn};
-
 use super::GpuBackend;
+use crate::core::air::accumulation::AccumulationOps;
+use crate::core::fields::secure_column::SecureColumn;
 
 impl AccumulationOps for GpuBackend {
     fn accumulate(_column: &mut SecureColumn<Self>, _other: &SecureColumn<Self>) {
