+++ crates/prover/src/core/backend/gpu/fri.rs
@@ -1,8 +1,9 @@
-use crate::core::{fields::qm31::SecureField, fri::FriOps, poly::{circle::SecureEvaluation, line::LineEvaluation, twiddles::TwiddleTree}};
-
 use super::GpuBackend;
-
-
+use crate::core::fields::qm31::SecureField;
+use crate::core::fri::FriOps;
+use crate::core::poly::circle::SecureEvaluation;
+use crate::core::poly::line::LineEvaluation;
+use crate::core::poly::twiddles::TwiddleTree;
 
 impl FriOps for GpuBackend {
     fn fold_line(
